Fix for issue 9720 and 9718 (#11)
* Fixes git hub issue 9720 and 9718
GH-9720 - The new load balancer code was doing a direct string match for determining two same hosts. This is not right specially for localhost and loopback addresses. Therefore this change is doing InetAddress equals now.

GH-9718 - The load balancer code while processing the new load balance properties like 'topology-keys' and 'load-balance' properties was not handling malformed url with empty values therefore the exception. Now the new behaviour is to log a warning and then fallback to original behaviour which connects to the database without using the new load balance behaviour.